requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.9.2
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-bootstrap==0.2.4
django-bootstrap4==25.2
django-bootstrap5==25.2
django-static==1.5.6
django-stdimage==6.0.2
frozenlist==1.8.0
geoip2==5.2.0
gunicorn==24.1.1
idna==3.11
maxminddb==3.0.0
multidict==6.7.0
packaging==26.0
pillow==11.3.0
propcache==0.4.1
psycopg2-binary==2.9.10
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
whitenoise==6.11.0
yarl==1.22.0
